Improved settings writing to be strongly typed and based on settings file content diffs
